+++ .lua-ls/wow-stdlib.lua
@@ -1,0 +1,9 @@
+---@meta
+
+---@class stringlib
+---@field lower fun(s: string): string
+---@field format fun(formatstring: string, ...: any): string
+---@field rep fun(s: string, n: integer): string
+
+---@type stringlib
+string = string